+++ app/Http/Controllers/SubjectController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Subject;
 use App\HospitalProfile;
 use App\SubjectJunctions;
+use DB;
 use Illuminate\Http\Request;
 
 class SubjectController extends Controller
@@ -134,10 +135,16 @@ class SubjectController extends Controller
     public function destroy($id)
     {
         $Subject = Subject::find($id);
-        $Subject->delete();
+        $sub_junction = "SELECT * FROM subject_junctions WHERE subject_junctions.subject_id = $id";
+        $junction_data = DB::select($sub_junction);
+        if(count($junction_data) != 0) {
+            return response()->json(['error' => 'この診療科目に関連している病院施設がありますので削除できません。'], 404);
+        }else{
+            $Subject->delete();
         // return response()->json('The Subject was successfully deleted');
         $subjects = Subject::orderBy('id', 'DESC')->paginate(12);
         return response()->json($subjects);
+        }        
     }
 
     public function search(Request $request) {
